Update path-to-regexp and implement pattern getters.
